requirements.txt
django==3.1.1
bunq_sdk==1.14.18
django-localflavor==3.0.1
django-money==1.1
currency-symbols==1.0.0